+++ lib/rails_performance/instrument/metrics_collector.rb
@@ -16,6 +16,7 @@ module RailsPerformance
       # }
 
       def call(event_name, started, finished, event_id, payload)
+        # TODO do we need this new?
         event = ActiveSupport::Notifications::Event.new(event_name, started, finished, event_id, payload)
 
         return if %r{#{RailsPerformance.mount_at}}.match? event.payload[:path]