+++ lib/rails_performance/rails/middleware.rb
@@ -1,5 +1,24 @@
 module RailsPerformance
   module Rails
+    class MiddlewareTraceStorerAndCleanup
+      def initialize(app)
+        @app = app
+      end
+
+      def call(env)
+        dup.call!(env)
+      end
+
+      def call!(env)
+        @status, @headers, @response = @app.call(env)
+
+        RP::Utils.log_trace_in_redis(CurrentRequest.current.request_id, CurrentRequest.current.storage)
+        CurrentRequest.cleanup
+
+        [@status, @headers, @response]
+      end
+    end
+
     class Middleware
       def initialize(app)
         @app = app
@@ -13,25 +32,19 @@ module RailsPerformance
         @status, @headers, @response = @app.call(env)
 
         #t = Time.now
-        if record = CurrentRequest.current.record
-          begin
-            record[:status]   ||= @status # for 500 errors
-            record[:request_id] = CurrentRequest.current.request_id
+        if !CurrentRequest.current.ignore.include?(:performance) && # grape is executed first, and than ignore regular future storage of "controller"-like request
+          record = CurrentRequest.current.record
 
-            # capture referer from where this page was opened
-            if record[:status] == 404
-              record[:HTTP_REFERER] = env["HTTP_REFERER"]
-            end
+          record[:status]   ||= @status # for 500 errors
+          record[:request_id] = CurrentRequest.current.request_id
 
-            # store for section "recent requests"
-            RP::Utils.log_trace_in_redis(CurrentRequest.current.request_id, CurrentRequest.current.storage)
+          # capture referer from where this page was opened
+          record[:HTTP_REFERER] = env["HTTP_REFERER"] if record[:status] == 404
 
-            # store request information
-            RP::Utils.log_request_in_redis(record)
-          ensure
-            # we don't want to have a memory leak
-            CurrentRequest.cleanup
-          end
+          # store for section "recent requests"
+
+          # store request information (regular rails request)
+          RP::Utils.log_request_in_redis(record)
         end
         #puts "==> store performance data: #{(Time.now - t).round(3)}ms"