+++ test/dummy/app/api/demo/worker2.rb
@@ -1,0 +1,9 @@
+module Demo
+  class Worker2 < Grape::API
+    desc 'Returns users.'
+    get :users do
+      sleep(rand(2.0))
+      User.limit(10).order("random()")
+    end
+  end
+end